clean up classes
