+++ src/test/java/uk/gov/ons/fsdr/tests/performance/utils/MockUtils.java
@@ -7,10 +7,8 @@ import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
-import org.springframework.web.util.UriComponentsBuilder;
 import uk.gov.ons.fsdr.common.dto.AdeccoResponse;
 import uk.gov.ons.fsdr.common.dto.AdeccoResponseList;
 import uk.gov.ons.fsdr.tests.performance.exceptions.MockInaccessibleException;
@@ -31,6 +29,15 @@ public final class MockUtils {
   @Value("${mock.baseUrl}")
   private String baseMockUrl;
 
+  @Value("${service.fsdrservice.url}")
+  private String fsdrServiceUrl;
+
+  @Value("${service.fsdrservice.username}")
+  private String fsdrServiceUsername;
+
+  @Value("${service.fsdrservice.password}")
+  private String fsdrServicePassword;
+
   @Value("${spring.datasource.url}")
   private String url;
 
@@ -39,6 +46,19 @@ public final class MockUtils {
 
   @Value("${spring.datasource.password}")
   private String password;
+
+  public void ingestAdecco() throws IOException {
+    URL url = new URL(fsdrServiceUrl + "/fsdr/adeccoIngest");
+    HttpURLConnection httpURLConnection = (HttpURLConnection) url.openConnection();
+    createBasicAuthHeaders(fsdrServiceUsername, fsdrServicePassword);
+
+    httpURLConnection.setRequestMethod("POST");
+    if (httpURLConnection.getResponseCode() != 200) {
+      log.error("failed to initiate Adecco ingest" + httpURLConnection.getResponseCode()
+          + httpURLConnection.getResponseMessage());
+      throw new RuntimeException(httpURLConnection.getResponseMessage());
+    }
+  }
 
   public void clearMock() throws IOException {
     URL url = new URL(baseMockUrl + "mock/reset");
@@ -59,12 +79,10 @@ public final class MockUtils {
   }
 
   public void clearDatabase() {
-    System.out.println("CLEARDB" + url + username + password);
     Statement stmt = null;
     try (Connection conn = DriverManager.getConnection(url, username, password)) {
-
       if (conn != null) {
-        System.out.println("Connected to the database!");
+        log.info("Connected to the database!");
         stmt = conn.createStatement();
         String sql = "DELETE FROM action_indicator";
         stmt.executeUpdate(sql);
@@ -86,7 +104,7 @@ public final class MockUtils {
         stmt.executeUpdate(sql);
 
       } else {
-        System.out.println("Failed to make connection!");
+        log.error("Failed to make connection!");
       }
     } catch (SQLException ignored) {
     } finally {
@@ -98,18 +116,9 @@ public final class MockUtils {
     }
   }
 
-  public AdeccoResponseList getRecords() {
-    RestTemplate restTemplate = new RestTemplate();
-    String url = baseMockUrl + "adecco/records";
-    log.info("getRecords-mock_url:" + url);
-    ResponseEntity<AdeccoResponseList> responseEntity;
-    responseEntity = restTemplate.getForEntity(url, AdeccoResponseList.class);
-    return responseEntity.getBody();
-  }
-
   public void addUsersAdecco(List<AdeccoResponse> adeccoResponseList) {
     RestTemplate restTemplate = new RestTemplate();
-    HttpHeaders headers = createBasicAuthHeaders("user", "password");
+    HttpHeaders headers = createBasicAuthHeaders(fsdrServiceUsername, fsdrServicePassword);
     headers.setContentType(MediaType.APPLICATION_JSON);
     HttpEntity<List<AdeccoResponse>> response = new HttpEntity<>(adeccoResponseList, headers);
     String postUrl = baseMockUrl + "mock/postResponse";
@@ -124,43 +133,5 @@ public final class MockUtils {
     String base64Creds = new String(base64CredsBytes);
     headers.add("Authorization", "Basic " + base64Creds);
     return headers;
-  }
-
-  public void enableRequestRecorder() throws IOException {
-    URL url = new URL(baseMockUrl + "mock/enable");
-    log.info("enableRequestRecorder-mock_url:" + url.toString());
-    HttpURLConnection httpURLConnection = (HttpURLConnection) url.openConnection();
-    httpURLConnection.setRequestMethod("GET");
-    if (httpURLConnection.getResponseCode() != 200) {
-      throw new MockInaccessibleException("Failed : HTTP error code : " + httpURLConnection.getResponseCode());
-    }
-  }
-
-  public void disableRequestRecorder() throws IOException {
-    URL url = new URL(baseMockUrl + "mock/disable");
-    log.info("disableRequestRecorder-mock_url:" + url.toString());
-    HttpURLConnection httpURLConnection = (HttpURLConnection) url.openConnection();
-    httpURLConnection.setRequestMethod("GET");
-    if (httpURLConnection.getResponseCode() != 200) {
-      throw new MockInaccessibleException("Failed : HTTP error code : " + httpURLConnection.getResponseCode());
-    }
-  }
-
-  public ResponseEntity<AdeccoResponseList> getEmployeeBySource(String source) {
-    RestTemplate restTemplate = new RestTemplate();
-    String postHit = baseMockUrl + "/getResponse";
-    ResponseEntity<AdeccoResponseList> results = restTemplate.exchange(postHit, HttpMethod.GET, null,
-        AdeccoResponseList.class);
-    return results;
-  }
-
-  public ResponseEntity<AdeccoResponseList> getEmployeeById(String employeeId) {
-    RestTemplate restTemplate = new RestTemplate();
-    String postHit = baseMockUrl + "fsdr/getEmployee/";
-    UriComponentsBuilder builder = UriComponentsBuilder.fromHttpUrl(postHit)
-        .queryParam("employeeId", employeeId);
-    ResponseEntity<AdeccoResponseList> results = restTemplate.exchange(builder.toUriString(), HttpMethod.GET, null,
-        AdeccoResponseList.class);
-    return results;
   }
 }
